Update plugins, add recording macros, add nowrap to Neotest
The biggest change here is the ability to run macros and edit them more
easily. We now can record with q and see the macro in the status line.
We can change registers with <leader>qt which switches from registers a,
b or c, which contain the macros. We can edit the macros with
<leader>qe, and we can run the currently selected macro with Q.